Custom Exception

+++ src/main/kotlin/com/github/wnuk/restomanager/service/impl/ReservationServiceImpl.kt
@@ -1,11 +1,10 @@
 package com.github.wnuk.restomanager.service.impl
 
 import com.github.wnuk.restomanager.dao.toReservationDto
-import com.github.wnuk.restomanager.dao.toUserDto
 import com.github.wnuk.restomanager.dto.ReservationDto
 import com.github.wnuk.restomanager.dto.toReservationDao
+import com.github.wnuk.restomanager.exception.CustomException
 import com.github.wnuk.restomanager.repository.ReservationRepository
-import com.github.wnuk.restomanager.repository.UserRepository
 import com.github.wnuk.restomanager.service.ReservationService
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
@@ -17,6 +16,10 @@ class ReservationServiceImpl(private val repository: ReservationRepository) : Re
     private val log = LoggerFactory.getLogger(ReservationServiceImpl::class.java)
 
     override fun deleteReservation(id: Long) {
+        val reservation = repository.findById(id)
+        if (reservation.isEmpty){
+            throw CustomException("Reservation Not Found")
+        }
         repository.deleteById(id)
     }
 
@@ -28,8 +31,12 @@ class ReservationServiceImpl(private val repository: ReservationRepository) : Re
     }
 
     override fun getReservationById(id: Long): ReservationDto? {
-        val reservation = repository.findById(id).get()
-        return reservation.toReservationDto()
+        val reservation = repository.findById(id)
+        if (reservation.isEmpty){
+            throw CustomException("Reservation Not Found")
+        }
+
+        return reservation.get().toReservationDto()
     }
 
     override fun getReservationByDate(from: Long, to: Long): List<ReservationDto>? {
@@ -53,6 +60,6 @@ class ReservationServiceImpl(private val repository: ReservationRepository) : Re
     override fun getAllReservations(): List<ReservationDto>? {
         return repository.findAll().map {
             it.toReservationDto()
-        }
+        }.ifEmpty { throw CustomException("Reservation Not Found") }
     }
 }
